refactor types.hpp into utils
