+++ src/attacks.hpp
@@ -22,7 +22,7 @@
 #include <cstdint>
 #include <vector>
 
-#include "types.hpp"
+#include "utils/types.hpp"
 
 class Attacks {
     public:
