+++ src/board.cpp
@@ -29,7 +29,8 @@
 #include "bitboard.hpp"
 #include "zobrist.hpp"
 #include "eval.hpp" 
-#include "types.hpp"
+
+#include "utils/types.hpp"
 
 Board::Board(std::string fenStr) {
     std::string token;
