+++ src/move.cpp
@@ -22,6 +22,8 @@
 #include <array>
 
 #include "move.hpp"
+#include "utils/enums.hpp"
+#include "utils/types.hpp"
 
 BoardMove::BoardMove(Square square1, Square square2, pieceTypes promotePiece) {
     data =  square1 | (square2 << 6) | (toInt(promotePiece) << 12);
