+++ src/moveOrder.cpp
@@ -23,7 +23,6 @@
 #include "moveOrder.hpp"
 #include "board.hpp"
 #include "move.hpp"
-#include "types.hpp"
 
 namespace MoveOrder {
